Split sync event hook bridge from runtime route
