+++ backend/threads/chat_adapters/runtime_chat_delivery_action.py
@@ -8,6 +8,7 @@ from backend.threads.chat_adapters.port import get_agent_runtime_gateway
 from backend.threads.chat_adapters.runtime_event_action_route import RuntimeEventActionRoute
 from backend.threads.chat_adapters.runtime_identity import runtime_actor
 from backend.threads.chat_adapters.runtime_recipient import resolve_runtime_chat_delivery_recipient
+from backend.threads.chat_adapters.runtime_sync_event_hook import make_sync_runtime_event_hook
 from protocols.agent_runtime import (
     AgentChatContext,
     AgentChatDeliveryEnvelope,
@@ -45,10 +46,11 @@ def make_runtime_chat_delivery_event_hook[EventT](
             activity_reader=activity_reader,
         )
 
-    return RuntimeEventActionRoute(
+    route = RuntimeEventActionRoute(
         planner=planner,
         dispatch_actions=dispatch_actions,
-    ).sync_hook()
+    )
+    return make_sync_runtime_event_hook(route.dispatch)
 
 
 async def dispatch_runtime_chat_delivery_actions(
