+++ backend/threads/chat_adapters/runtime_event_action_route.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import asyncio
 from collections.abc import Callable, Coroutine, Iterable
 from dataclasses import dataclass
 from typing import Any
@@ -13,21 +12,3 @@ class RuntimeEventActionRoute[EventT, ActionT, ResultT]:
 
     async def dispatch(self, event: EventT) -> ResultT:
         return await self.dispatch_actions(self.planner(event))
-
-    def sync_hook(self) -> Callable[[EventT], None]:
-        loop = asyncio.get_running_loop()
-
-        async def dispatch_event(event: EventT) -> None:
-            await self.dispatch(event)
-
-        def hook(event: EventT) -> None:
-            try:
-                current_loop = asyncio.get_running_loop()
-            except RuntimeError:
-                current_loop = None
-            if current_loop is loop:
-                raise RuntimeError("Sync runtime event hook cannot run on its owner event loop thread")
-            future = asyncio.run_coroutine_threadsafe(dispatch_event(event), loop)
-            future.result()
-
-        return hook
