+++ backend/threads/chat_adapters/runtime_notification_action.py
@@ -8,6 +8,7 @@ from backend.threads.chat_adapters.port import get_agent_runtime_gateway
 from backend.threads.chat_adapters.runtime_event_action_route import RuntimeEventActionRoute
 from backend.threads.chat_adapters.runtime_identity import make_runtime_actor, require_user
 from backend.threads.chat_adapters.runtime_recipient import resolve_runtime_notification_recipient
+from backend.threads.chat_adapters.runtime_sync_event_hook import make_sync_runtime_event_hook
 from protocols.agent_runtime import (
     AgentRuntimeMessage,
     AgentRuntimeNotificationEnvelope,
@@ -47,10 +48,11 @@ def make_runtime_notification_event_hook[EventT](
             activity_reader=activity_reader,
         )
 
-    return RuntimeEventActionRoute(
+    route = RuntimeEventActionRoute(
         planner=planner,
         dispatch_actions=dispatch_actions,
-    ).sync_hook()
+    )
+    return make_sync_runtime_event_hook(route.dispatch)
 
 
 async def dispatch_runtime_notification_actions(
