+++ backend/threads/chat_adapters/runtime_sync_event_hook.py
@@ -1,0 +1,23 @@
+from __future__ import annotations
+
+import asyncio
+from collections.abc import Callable, Coroutine
+from typing import Any
+
+
+def make_sync_runtime_event_hook[EventT](
+    dispatch_event: Callable[[EventT], Coroutine[Any, Any, Any]],
+) -> Callable[[EventT], None]:
+    loop = asyncio.get_running_loop()
+
+    def hook(event: EventT) -> None:
+        try:
+            current_loop = asyncio.get_running_loop()
+        except RuntimeError:
+            current_loop = None
+        if current_loop is loop:
+            raise RuntimeError("Sync runtime event hook cannot run on its owner event loop thread")
+        future = asyncio.run_coroutine_threadsafe(dispatch_event(event), loop)
+        future.result()
+
+    return hook
